port and refactor rolls
