+++ src/MyMonster.js
@@ -1,48 +1,5 @@
 import { useState, useEffect } from "react";
 
-export function roll(string = "1d20") {
-    // parse string
-    console.log("string", string)
-    let dPos = 0;
-    let modPos = string.length;
-    let count = 0;
-    let diceType = 0;
-    let mod = 0;
-    for(let i=0; i<string.length; i++) {
-        // console.log(string[i]);
-        if(string[i]==='d') {
-            dPos = i;
-        } else if(string[i]==='+' || string[i]==='-') {
-            modPos = i;
-        }
-    }
-    count = Number(string.slice(0,dPos));
-    diceType = Number(string.slice(dPos+1,modPos));
-    mod = Number(string.slice(modPos));
-    // console.log("count",count);
-    // console.log("d", diceType);
-    // console.log("mod", mod);
-
-    // roll
-    let result = 0;
-    let roll = 0
-    for(let i=0; i<count; i++) {
-        roll = Math.floor(Math.random() * diceType) + 1;
-        // console.log("roll", i, roll);
-        result += roll;
-    }
-    result += mod;
-    console.log("result", result);
-    return result;
-}
-
-function rollAbility(ability) {
-    let mod = calcAbilityMod(ability);
-    let roll = Math.floor(Math.random() * 20) + 1;
-    console.log("result", (roll+mod));
-    return roll+mod;
-}
-
 function calcAbilityMod(num) {
     return (Math.floor((num-10)/2))
 }
@@ -94,27 +51,27 @@ export default function MyMonster({ monster, onRoll, onDelete }) {
                     <hr />
                     {/* Abilities */}
                     <div className="row">
-                        <div className="col text-center" onClick={() => onRoll("STR", "1d20+mod", rollAbility(monster.strength))}>
+                        <div className="col text-center" onClick={() => onRoll("STR", `1d20+${modifiers.strength}`)}>
                             <div>STR</div>
                             <div>{monster.strength} ({modifiers.strength})</div>
                         </div>
-                        <div className="col text-center" onClick={() => onRoll("DEX", "1d20+mod", rollAbility(monster.dexterity))}>
+                        <div className="col text-center" onClick={() => onRoll("DEX", `1d20+${modifiers.dexterity}`)}>
                             <div>DEX</div>
                             <div>{monster.dexterity} ({modifiers.dexterity})</div>
                         </div>
-                        <div className="col text-center" onClick={() => onRoll("CON", "1d20+mod", rollAbility(monster.constitution))}>
+                        <div className="col text-center" onClick={() => onRoll("CON", `1d20+${modifiers.constitution}`)}>
                             <div>CON</div>
                             <div>{monster.constitution} ({modifiers.constitution})</div>
                         </div>
-                        <div className="col text-center" onClick={() => onRoll("INT", "1d20+mod", rollAbility(monster.intelligence))}>
+                        <div className="col text-center" onClick={() => onRoll("INT", `1d20+${modifiers.intelligence}`)}>
                             <div>INT</div>
                             <div>{monster.intelligence} ({modifiers.intelligence})</div>
                         </div>
-                        <div className="col text-center" onClick={() => onRoll("WIS", "1d20+mod", rollAbility(monster.wisdom))}>
+                        <div className="col text-center" onClick={() => onRoll("WIS", `1d20+${modifiers.wisdom}`)}>
                             <div>WIS</div>
                             <div>{monster.wisdom} ({modifiers.wisdom})</div>
                         </div>
-                        <div className="col text-center" onClick={() => onRoll("CHA", "1d20+mod", rollAbility(monster.charisma))}>
+                        <div className="col text-center" onClick={() => onRoll("CHA", `1d20+${modifiers.charisma}`)}>
                             <div>CHA</div>
                             <div>{monster.charisma} ({modifiers.charisma})</div>
                         </div>
@@ -127,13 +84,14 @@ export default function MyMonster({ monster, onRoll, onDelete }) {
                 </div>
                 <div className="col">
                     {monster.special_abilities ? monster.special_abilities.map((sa, index) => (
-                    <div key={index}><span style={{fontWeight:"bold"}}>{sa.name}{sa.usage && <> ({sa.usage.type})</>}. </span>{sa.desc}</div>
+                    <div key={index} onClick={() => {if(sa.damage_dice){onRoll(sa.name,`${sa.damage_dice}+${sa.damage_bonus ? sa.damage_bonus : 0}`)}}}>
+                        <span style={{fontWeight:"bold"}}>{sa.name}. </span>{sa.desc}</div>
                     )) : null}
                     <hr />
                     <h5>Actions</h5>
                     {monster.actions.map((action, index) => (
-                        <div key={index} onClick={() => {if(action.damage){onRoll(action.name,action.damage[0].damage_dice,roll(action.damage[0].damage_dice))}}}>
-                            <span style={{fontWeight:"bold"}}>{action.name}{action.usage && <> ({action.usage.type})</>}. </span>{action.desc}
+                        <div key={index} onClick={() => {if(action.damage_dice){onRoll(action.name,`${action.damage_dice}+${action.damage_bonus ? action.damage_bonus : 0}`)}}}>
+                            <span style={{fontWeight:"bold"}}>{action.name}. </span>{action.desc}
                         </div>
                     ))}
                 </div>
