+++ src/log.js
@@ -1,8 +1,9 @@
-export default class Roll {
-    constructor(creature,type,dice,result) {
+import { roll } from "./die";
+export default class LogRoll {
+    constructor(creature,type,dice) {
         this.creature = creature;
         this.type = type;
         this.dice = dice;
-        this.result = result;
+        this.result = roll(dice);
     }
 }
